requirements/dev.txt
# Development & Testing requirements
# Include base production requirements
-r base.txt

# Testing frameworks
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
coverage==7.4.3

# Test data factories
factory-boy==3.3.0
faker==24.0.0
model-bakery==1.17.0

# Development tools
django-debug-toolbar==6.1.0
django-extensions==4.1
